fix building library shell

+++ scripts/build_lib.sh
@@ -1,2 +1,6 @@
-rm -r dist
+if [ -d "./dist/lib" ]
+then
+  echo "has dist"
+  rm -r dist
+fi
 ng build lib --watch
